+++ jdaviz/configs/default/plugins/export/export.vue
@@ -1,5 +1,8 @@
 <template>
   <j-tray-plugin
+    :config="config"
+    plugin_key="Export"
+    :api_hints_enabled.sync="api_hints_enabled"
     description='Export data or plots from the app to a file.'
     :link="docs_link || 'https://jdaviz.readthedocs.io/en/'+vdocs+'/'+config+'/plugins.html#export'"
     :popout_button="popout_button"
@@ -22,16 +25,17 @@
         :selected.sync="viewer_selected"
         :multiselect="multiselect"
         :single_select_allow_blank="false"
+        api_hint="plg.viewer ="
+        :api_hints_enabled="api_hints_enabled"
       >
       </plugin-inline-select>
       <v-row class="row-min-bottom-padding">
         <v-select
-          class="category-content"
-          :menu-props="{ left: true }"
           attach
           v-model="viewer_format_selected"
           :items="viewer_format_items.map(i => i.label)"
-          label="Format"
+          :label="api_hints_enabled ? 'plg.viewer_format =' : 'Format'"
+          :class="api_hints_enabled ? 'api-hint category-content' : 'category-content'"
           hint="Image format for exporting viewers."
           :disabled="viewer_selected.length == 0"
           persistent-hint
@@ -101,6 +105,8 @@
         :selected.sync="dataset_selected"
         :multiselect="multiselect"
         :single_select_allow_blank="false"
+        api_hint="plg.dataset ="
+        :api_hints_enabled="api_hints_enabled"
       >
       </plugin-inline-select>
 
@@ -111,12 +117,11 @@
       </v-row>
       <v-row class="row-min-bottom-padding">
           <v-select
-            class="category-content"
-            :menu-props="{ left: true }"
             attach
             v-model="dataset_format_selected"
             :items="dataset_format_items.map(i => i.label)"
-            label="Format"
+            :label="api_hints_enabled ? 'plg.dataset_format =' : 'Format'"
+            :class="api_hints_enabled ? 'api-hint category-content' : 'category-content'"
             hint="Format for exporting datasets."
             :disabled="dataset_selected.length == 0"
             persistent-hint
@@ -135,6 +140,8 @@
         :selected.sync="subset_selected"
         :multiselect="multiselect"
         :single_select_allow_blank="false"
+        api_hint="plg.subset ="
+        :api_hints_enabled="api_hints_enabled"
       >
       </plugin-inline-select>
 
@@ -146,14 +153,13 @@
 
       <v-row class="row-min-bottom-padding">
         <v-select
-          class="category-content"
-          :menu-props="{ left: true }"
           attach
           v-model="subset_format_selected"
           :items="subset_format_items"
           item-text="label"
           item-disabled="disabled"
-          label="Format"
+          :label="api_hints_enabled ? 'plg.subset_format =' : 'Format'"
+          :class="api_hints_enabled ? 'api-hint category-content' : 'category-content'"
           hint="Format for exporting subsets."
           :disabled="subset_selected == null || subset_selected.length == 0"
           persistent-hint
@@ -172,16 +178,17 @@
         :selected.sync="plugin_table_selected"
         :multiselect="multiselect"
         :single_select_allow_blank="false"
+        api_hint="plg.table ="
+        :api_hints_enabled="api_hints_enabled"
       >
       </plugin-inline-select>
       <v-row class="row-min-bottom-padding">
         <v-select
-          class="category-content"
-          :menu-props="{ left: true }"
           attach
           v-model="plugin_table_format_selected"
           :items="plugin_table_format_items.map(i => i.label)"
-          label="Format"
+          :label="api_hints_enabled ? 'plg.plugin_table_format =' : 'Format'"
+          :class="api_hints_enabled ? 'api-hint category-content' : 'category-content'"
           hint="File format for exporting plugin tables."
           :disabled="plugin_table_selected.length == 0"
           persistent-hint
@@ -200,6 +207,8 @@
         :selected.sync="plugin_plot_selected"
         :multiselect="multiselect"
         :single_select_allow_blank="false"
+        api_hint="plg.plugin_plot ="
+        :api_hints_enabled="api_hints_enabled"
       >
       </plugin-inline-select>
       <jupyter-widget
@@ -208,12 +217,11 @@
           :widget="plugin_plot_selected_widget"/>
       <v-row class="row-min-bottom-padding">
         <v-select
-          class="category-content"
-          :menu-props="{ left: true }"
           attach
           v-model="plugin_plot_format_selected"
           :items="plugin_plot_format_items.map(i => i.label)"
-          label="Format"
+          :label="api_hints_enabled ? 'plg.plugin_plot_format =' : 'Format'"
+          :class="api_hints_enabled ? 'api-hint category-content' : 'category-content'"
           hint="File format for exporting plugin plots."
           :disabled="plugin_plot_selected.length == 0"
           persistent-hint
@@ -243,6 +251,8 @@
       :auto.sync="filename_auto"
       :invalid_msg="filename_invalid_msg"
       label="Filename"
+      api_hint="plg.filename ="
+      :api_hints_enabled="api_hints_enabled"
       hint="Export to a file on disk."
     ></plugin-auto-label>
 
@@ -261,13 +271,18 @@
         :results_isolated_to_plugin="true"
         @click="export_from_ui"
         :spinner="spinner"
+        :api_hints_enabled="api_hints_enabled"
         :disabled="filename_value.length === 0 ||
                    movie_recording ||
                    subset_invalid_msg.length > 0 || data_invalid_msg.length > 0 ||
                    viewer_invalid_msg.length > 0 ||
                    (viewer_selected.length > 0 && viewer_format_selected == 'mp4' && !movie_enabled)"
       >
-        Export
+        {{ api_hints_enabled ?
+          'plg.export()'
+          :
+          'Export'
+        }}
       </plugin-action-button>
     </div>