+++ jdaviz/configs/specviz/plugins/unit_conversion/unit_conversion.vue
@@ -1,6 +1,8 @@
 <template>
-  <!-- To re-enable plugin, use :disabled_msg="disabled_msg" -->
   <j-tray-plugin
+    :config="config"
+    plugin_key="Unit Conversion"
+    :api_hints_enabled.sync="api_hints_enabled"
     :description="docs_description || 'Convert the spectral flux density and spectral axis units.'"
     :link="docs_link || 'https://jdaviz.readthedocs.io/en/'+vdocs+'/'+config+'/plugins.html#unit-conversion'"
     :disabled_msg="disabled_msg"
@@ -13,7 +15,8 @@
         attach
         :items="spectral_unit_items.map(i => i.label)"
         v-model="spectral_unit_selected"
-        label="Spectral Unit"
+        :label="api_hints_enabled ? 'plg.spectral_unit =' : 'Spectral Unit'"
+        :class="api_hints_enabled ? 'api-hint' : null"
         hint="Global display unit for spectral axis."
         persistent-hint
       ></v-select>
@@ -25,7 +28,8 @@
         attach
         :items="flux_unit_items.map(i => i.label)"
         v-model="flux_unit_selected"
-        label="Flux Unit"
+        :label="api_hints_enabled ? 'plg.flux_unit =' : 'Flux Unit'"
+        :class="api_hints_enabled ? 'api-hint' : null"
         hint="Global display unit for flux axis."
         persistent-hint
       ></v-select>
@@ -37,7 +41,8 @@
         attach
         :items="angle_unit_items.map(i => i.label)"
         v-model="angle_unit_selected"
-        label="Solid Angle Unit"
+        :label="api_hints_enabled ? 'plg.angle_unit =' : 'Solid Angle Unit'"
+        :class="api_hints_enabled ? 'api-hint' : null"
         hint="Solid angle unit."
         persistent-hint
         
@@ -47,7 +52,8 @@
     <v-row>
       <v-text-field
         v-model="sb_unit_selected"
-        label="Surface Brightness Unit"
+        :label="api_hints_enabled ? 'plg.sb_unit' : 'Surface Brightness Unit'"
+        :class="api_hints_enabled ? 'api-hint' : null"
         hint="Global display unit for surface brightness axis."
         persistent-hint
         :disabled='true'
@@ -64,7 +70,8 @@
         attach
         :items="flux_or_sb_items.map(i => i.label)"
         v-model="flux_or_sb_selected"
-        label="Flux or Surface Brightness"
+        :label="api_hints_enabled ? 'plg.flux_or_sb =' : 'Flux or Surface Brightness'"
+        :class="api_hints_enabled ? 'api-hint' : null"
         hint="Select the y-axis physical type for the spectrum-viewer."
         persistent-hint
         :disabled="!can_translate"