+++ jdaviz/main_styles.vue
@@ -262,4 +262,41 @@ a:active {
   font-family: Roboto, sans-serif !important;
   font-weight: 500 !important;
 }
+
+span.api-hint, span.api-hint-header {
+  color: #c75109 !important;
+  font-family: monospace;
+  font-size: 12px;  
+}
+
+span.api-hint-header {
+  font-weight: bold;
+}
+
+.api-hint :is(.v-select__slot, .v-text-field__slot, .v-input__slot, .v-input--radio-group__input) > .v-label {
+    color: #c75109 !important;
+    font-family: monospace;
+    font-size: 10px !important;
+    overflow-wrap: break-word;
+}
+
+.api-hint :is(.v-select__slot, .v-text-field__slot) > .v-label {
+    font-size: 13px !important;
+}
+
+
+.api-hint .v-alert__content, .api-hint.slider-label {
+  color: #c75109 !important;
+  font-family: monospace;
+  font-size: 10px !important;
+}
+
+.api-hint.v-btn {
+  color: #c75109 !important;
+  font-family: monospace !important;
+  font-size: 12px !important;
+  overflow-wrap: break-word;
+  text-transform: none !important;
+}
+
 </style>